Refactor: inject packageInfo instead of importing it
- Server now accepts packageInfo as an injected dependency via ServerOptions
- CLI factory function createServerFromUserConfig now accepts and uses packageInfo
- ServerFactory simplified to only create infrastructure (config, logger, metrics, keychain)
- mongodb-mcp-server index.ts now creates Server with all dependencies including packageInfo
- PackageInfo type added to CLI exports
- Removed incomplete handler files from CLI package
- Removed duplicate server.ts from CLI package

This allows CLI to remain independent of mongodb-mcp-server while still
providing factory functions for shared infrastructure.
